+++ example/38_grouped_conv_bwd_data_multiple_d/common_conv3d.hpp
