+++ example/38_grouped_conv_bwd_data_multiple_d/grouped_conv3d_bwd_data_wmma_v3_bf16.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "ck/tensor_operation/gpu/device/impl/device_grouped_conv_bwd_data_multiple_d_wmma_cshuffle_v3.hpp"
-#include "common_3dtests.hpp"
+#include "common_conv3d.hpp"
 
 using OutDataType      = BF16;
 using WeiDataType      = BF16;
